+++ www/modules/email/EmailComposer.js
@@ -396,7 +396,6 @@ GO.email.EmailComposer = function(config) {
 	tbar.push(this.emailEditor.getAttachmentsButton(), 
 			this.showMenuButton = new Ext.Button({
 				tooltip : t("Add recipients", "email"),
-				// iconCls : 'ic-contact-mail',
 				iconCls : 'ic-people',
 				menu : this.showMenu
 			}));
@@ -511,8 +510,6 @@ GO.email.EmailComposer = function(config) {
 									template_id: "default_for_account"
 								});
 
-								
-								
 								this.fireEvent('load', this, records);
 								this.loaded = true;
 							}
@@ -542,11 +539,6 @@ GO.email.EmailComposer = function(config) {
 								var fromComboValue = this.fromCombo.getValue();
 								this.fromCombo.store.load();
 								this.fromCombo.setValue(fromComboValue);
-								
-								
-								
-								
-								
 							},
 							scope : this
 						}),
@@ -1233,16 +1225,6 @@ Ext.extend(GO.email.EmailComposer, GO.Window, {
 		}
 	},
 	
-
-	HandleResult : function (btn){
-		if (btn == 'yes'){
-			//this.htmlEditor.SpellCheck = true;
-			this.sendMail();
-		}else{
-			//this.editor.plugins[1].spellcheck();
-		}
-	},
-
 	submitForm : function(hide){
 		this.sendMail(false, false);
 	},